Revert "resolve Z202 Found too many module members"
This reverts commit 2ad986f71181deef1ff779aa3f02638c8f8ebc7c.
